+++ microservice-pos/src/main/java/com/cumlpeanos/pos/service/ICajaPOSService.java
@@ -3,5 +3,5 @@ package com.cumlpeanos.pos.service;
 import com.cumlpeanos.pos.models.entity.CajaPOS;
 
 public interface ICajaPOSService {
-    CajaPOS findByAlmacenAndPventa(int almacen, int pventa);
+    CajaPOS findByAlmacenAndPventa(Long almacen, Long pventa);
 }
